requirements.txt
streamlit
pandas
plotly
openpyxl
kaleido
python-dotenv
Pillow
matplotlib
